Refactored guides
