+++ apps/crn-server/src/controllers/guide.controller.ts
@@ -1,14 +1,10 @@
 import { ListGuideResponse } from '@asap-hub/model';
 import { GuideDataProvider } from '../data-providers/types';
 
-export default class Guide implements GuideController {
+export default class GuideController {
   constructor(private guideDataProvider: GuideDataProvider) {}
 
   async fetch(): Promise<ListGuideResponse> {
     return this.guideDataProvider.fetch();
   }
-}
-
-export interface GuideController {
-  fetch: () => Promise<ListGuideResponse>;
 }
