+++ apps/crn-server/src/routes/guide.route.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { GuideController } from '../controllers/guides.controller';
+import GuideController from '../controllers/guide.controller';
 
 export const guideRouteFactory = (guideController: GuideController): Router => {
   const guideRoutes = Router();
